requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Django==3.0.5
django-filter==2.2.0
djangorestframework==3.11.0
idna==2.9
pkg-resources==0.0.0
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
